fix api cell (add storage_id)

+++ app/Http/Controllers/Api/CellController.php
@@ -14,79 +14,53 @@ class CellController extends Controller
 {
     public function CellUpdate(Request $request, JSONcontroller $JSON)
     {
-
         try {
-//            $vId = $request->get('id');
-//            $vRack = $request->get('rack');
-//            $vStorey = $request->get('storey');
-//            $vRow = $request->get('row');
-            //$vId = $request->post('id');
             $vRack = $request->post('rack');
             $vStorey = $request->post('storey');
             $vRow = $request->post('row');
-            if ($vRack == null || $vStorey == null || $vRow == null) {
-                return $JSON->JSONerror('Нічого не передали або немає аргументів `rack`,`storey`,`row`', 401);
+            $vStorage = $request->post('storage');
+            if ($vRack == null || $vStorey == null || $vRow == null || $vStorage == null) {
+                return $JSON->JSONerror('Нічого не передали або немає аргументів `rack`,`storey`,`row`, `storage``', 401);
             }
-//            $vLastCell = DB::table('cell')->where('rack', $vRack)->latest('storey')->latest('row')->first();
-//            $vLastStorey = $vLastCell->storey;
-//            $vLastRow = $vLastCell->row;
-
-//            $vAllCell = DB::table('cell')->where('rack',$vRack)->latest('storey')->latest('row')->get();
-            $vAllCell = DB::table('cell')->where('rack', $vRack)->latest('storey')->latest('row')->get();
+            $vAllCell = DB::table('cell')->where('rack', $vRack)->where('storage_id', $vStorage)->latest('storey')->latest('row')->get();
             foreach ($vAllCell as $cell) {
                 if ($cell->storey > $vStorey)
                     Cell::destroy($cell->id);
                 if ($cell->row > $vRow)
                     Cell::destroy($cell->id);
-                //dd($cell);
             }
             for ($s = 1; $s <= $vStorey; $s++) {
                 for ($r = 1; $r <= $vRow; $r++) {
-                    if (DB::table('cell')->where('rack', $vRack)->where('storey', $s)->where('row', $r)->get()->first() == null) {
+                    if (DB::table('cell')->where('rack', $vRack)->where('storage_id', $vStorage)->where('storey', $s)->where('row', $r)->get()->first() == null) {
                         $cell = new Cell();
                         $cell->rack = $vRack;
                         $cell->storey = $s;
                         $cell->row = $r;
+                        $cell->storage_id = $vStorage;
                         $cell->save();
                     }
                 }
             }
-            $cellAll = DB::table('cell')->where('rack', $vRack)->latest('storey')->latest('row')->get();
+            $cellAll = DB::table('cell')->where('rack', $vRack)->where('storage_id',$vStorage)->latest('storey')->latest('row')->get();
             return $JSON->JSONsuccessArray('Update', 'New Cells', $cellAll, 201);
 
         } catch (\Exception $e) {
             return $JSON->JSONerror($e->getMessage(), 501);
         }
-//        try {
-//            $cell = new Cell();
-//            $cell->rack = $request->input('rack');
-//            $cell->storey = $request->input('storey');
-//            $cell->row = $request->input('row');
-//            $cell->save();
-//            //$cell->transform_cell($request);
-//            return $JSON->JSONsuccess("Успіх", 201);
-//
-//        } catch (\Exception $e) {
-//            return $JSON->JSONerror($e->getMessage(), 401);
-//        }
-
-
     }
 
     public function CellDestroy(Request $request, JSONcontroller $JSON)
     {
         try {
-//            $vId = $request->get('id');
-//            $vRack = $request->get('rack');
             $vId = $request->post('id');
             $vRack = $request->post('rack');
-            $vStorage = $request->post('storage_id');
+            $vStorage = $request->post('storage');
 
             if ($vId != null) {
                 Cell::destroy($vId);
                 return $JSON->JSONsuccess('Destroy element by id = ' . $vId, 201);
             } else if ($vRack != null) {
-                $vAllCell = Cell::whereRack($request->post('rack'));
+                $vAllCell = Cell::whereRack($request->post('rack'))->whereStorage_id($vStorage);
                 $count = $vAllCell->delete();
 //                $vAllCell = DB::table('cell')->where('rack', $vRack)->get();
 //                $count = 0;
@@ -104,34 +78,26 @@ class CellController extends Controller
     }
 
 
-    public function CellGet(Request $request, JSONcontroller $JSON)
+    public function cellGet(Request $request, JSONcontroller $JSON)
     {
         try {
-            if ($vId = $request->get('id') != null) {
+            $vRack = $request->get('rack');
+            $vId = $request->get('id');
+            $vStorage = $request->get('storage');
 
-                $vCall = DB::table('cell')->where('id', $vId)->get();
-                return $JSON->JSONsuccessArray('get by id: `' . $vId . '`', 'Cells', $vCall, 200);
-            } else if (($vRack = $request->get('rack')) != null) {
-                $vCall = DB::table('cell')->where('rack', $vRack)->get();
-                return $JSON->JSONsuccessArray('get by rack: `' . $vRack . '`', 'Cells', $vCall, 200);
-            } else {
-                return $JSON->JSONsuccessArray('all', 'Cells', Cell::all(), 200);
+            if($vId != null)
+            {
+             $vcell = Cell::whereId($vId)->get();
+                return $JSON->JSONsuccessArray('get by id: `' . $vId . '`', 'Cells', $vcell, 200);
             }
-//            $Cell = new Cell();
-//            try {
-//                $request->validate([
-//                    'rack' => 'required'
-//                ]);
-//            } catch (\Exception) {
-//                return $JSON->JSONsuccessArray('Get  all', 'Cell', $Cell::all(), 200);
-//            }
-//
-//            /*виводимо з бази по фільтру*/
-//            return $JSON->JSONsuccessArray(
-//                'Get rack by `' . $request->input('rack') . '`',
-//                'Cell',
-//                $Cell::where('rack', $request->input('rack'))->get(),
-//                200);
+            if($vStorage == null){
+                return $JSON->JSONerror('`storage` field is missing', 400);
+            }
+            if($vRack == null){
+                return $JSON->JSONsuccessArray('all', 'Cells', Cell::whereStorage_id($vStorage)->get(), 200);
+            }else{
+                return $JSON->JSONsuccessArray('all', 'Cells', Cell::whereStorage_id($vStorage)->whereRack($vRack)->get(), 200);
+            }
 
         } catch (\Exception $e) {
             return $JSON->JSONerror($e->getMessage(), 400);
@@ -180,33 +146,33 @@ class CellController extends Controller
      */
 
 
-    public
-    function set_cell2(Request $request, JSONcontroller $JSON)
-    {
-        try {
-            try {
-                $request->validate([
-                    'rack' => 'required',
-                    'storey' => 'required',
-                    'row' => 'required'
-                ]);
-            } catch (\Exception) {
-                throw new \Exception('У запиті не знайшло `rack, storey або row`');
-            }
-
-//           $gg = DB::select('SELECT * FROM `cell`');
-//            dd($gg);
-            $review = new Cell();
-            $review->rack = $request->input('rack');
-            $review->storey = $request->input('storey');
-            $review->row = $request->input('row');
-            $review->save();
-
-            return $JSON->JSONsuccess('Поле успішно додано.', 201);
-
-
-        } catch (\Exception $e) {
-            return $JSON->JSONerror($e->getMessage(), 400);
-        }
-    }
+//    public
+//    function set_cell2(Request $request, JSONcontroller $JSON)
+//    {
+//        try {
+//            try {
+//                $request->validate([
+//                    'rack' => 'required',
+//                    'storey' => 'required',
+//                    'row' => 'required'
+//                ]);
+//            } catch (\Exception) {
+//                throw new \Exception('У запиті не знайшло `rack, storey або row`');
+//            }
+//
+////           $gg = DB::select('SELECT * FROM `cell`');
+////            dd($gg);
+//            $review = new Cell();
+//            $review->rack = $request->input('rack');
+//            $review->storey = $request->input('storey');
+//            $review->row = $request->input('row');
+//            $review->save();
+//
+//            return $JSON->JSONsuccess('Поле успішно додано.', 201);
+//
+//
+//        } catch (\Exception $e) {
+//            return $JSON->JSONerror($e->getMessage(), 400);
+//        }
+//    }
 }
